refactor:closet 폴더구조 개선
closet 폴더구조 개선

#48
